moved date formatter to `CodeEditUtils`
