+++ CodeEdit/Info.plist
@@ -45,6 +45,6 @@
 		</dict>
 	</array>
 	<key>GitHash</key>
-	<string>6497e2f5a4108929a882923308d5817d44c72697</string>
+	<string>962d8a098f75c00ffec670a8b4c61230605e498d</string>
 </dict>
 </plist>
