+++ CodeEdit/InspectorSidebar/Views/HistoryItem.swift
@@ -6,28 +6,7 @@
 //
 import SwiftUI
 import GitClient
-
-extension Date {
-    func relativeStringToNow() -> String {
-        if Calendar.current.isDateInToday(self) ||
-            Calendar.current.isDateInYesterday(self) {
-            let style = RelativeFormatStyle(
-                presentation: .named,
-                unitsStyle: .abbreviated,
-                locale: .current,
-                calendar: .current,
-                capitalizationContext: .standalone
-            )
-
-            return self.formatted(style)
-        }
-        let formatter = DateFormatter()
-        formatter.dateStyle = .short
-        formatter.timeStyle = .none
-
-        return formatter.string(from: self)
-    }
-}
+import CodeEditUtils
 
 struct HistoryItem: View {
